# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RecoveryOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patient/chat:
    get:
      tags: [Chat]
      operationId: listMyChatMessages
      summary: List chat history with the AI assistant
      responses:
        '200':
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Chat]
      operationId: sendMyChatMessage
      summary: Send a message to the AI assistant and get a reply
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendChatMessageBody'
      responses:
        '200':
          description: Latest user + assistant pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatExchange'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          description: Free tier message limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /admin/dashboard:
    get:
      tags: [Admin]
      operationId: getAdminDashboard
      summary: Admin overview of patients, adherence, and pain
      responses:
        '200':
          description: Admin dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/patients:
    get:
      tags: [Admin]
      operationId: listAdminPatients
      summary: List all patients (admin)
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: condition
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminPatientRow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/patients/{patientId}:
    get:
      tags: [Admin]
      operationId: getAdminPatient
      summary: Get a single patient's full record (admin)
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPatientDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/patients/{patientId}/plan:
    put:
      tags: [Admin]
      operationId: updateAdminPatientPlan
      summary: Manually edit a patient's recovery plan (admin)
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryPlanContent'
      responses:
        '200':
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/patients/{patientId}/notes:
    get:
      tags: [Admin]
      operationId: listAdminPatientNotes
      summary: List clinician notes for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClinicianNote'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin]
      operationId: createAdminPatientNote
      summary: Add a clinician note for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClinicianNoteBody'
      responses:
        '200':
          description: Created note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicianNote'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PatientProfile:
      type: object
      required: [id, userId, age, gender, conditionId, conditionLabel, symptoms, painLevel, medicalHistory, reportFileName, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        age:
          type: integer
        gender:
          type: string
        conditionId:
          type: string
        conditionLabel:
          type: string
        symptoms:
          type: string
        painLevel:
          type: integer
        medicalHistory:
          type: string
        reportFileName:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Exercise:
      type: object
      required: [name, sets, reps, durationMinutes, instructions]
      properties:
        name:
          type: string
        sets:
          type: integer
        reps:
          type: integer
        durationMinutes:
          type: integer
        instructions:
          type: string
    WeeklyDay:
      type: object
      required: [day, focus, activities]
      properties:
        day:
          type: string
        focus:
          type: string
        activities:
          type: array
          items:
            type: string
    RecoveryPlanContent:
      type: object
      required: [summary, exercises, lifestyleTips, precautions, weeklyPlan]
      properties:
        summary:
          type: string
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/Exercise'
        lifestyleTips:
          type: array
          items:
            type: string
        precautions:
          type: array
          items:
            type: string
        weeklyPlan:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyDay'
    RecoveryPlan:
      type: object
      required: [id, patientId, content, generatedBy, createdAt, updatedAt]
      properties:
        id:
          type: string
        patientId:
          type: string
        content:
          $ref: '#/components/schemas/RecoveryPlanContent'
        generatedBy:
          type: string
          enum: [ai, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TrackingEntry:
      type: object
      required: [id, patientId, date, painScore, completedExercises, totalExercises, adherencePercent, notes, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: string
        date:
          type: string
          format: date
        painScore:
          type: integer
        completedExercises:
          type: integer
        totalExercises:
          type: integer
        adherencePercent:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    SendChatMessageBody:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
    ChatExchange:
      type: object
      required: [userMessage, assistantMessage]
      properties:
        userMessage:
          $ref: '#/components/schemas/ChatMessage'
        assistantMessage:
          $ref: '#/components/schemas/ChatMessage'
...
    AdminPatientRow:
      type: object
      required: [patientId, userId, displayName, email, conditionLabel, painLevel, latestPainScore, latestAdherence, lastCheckinAt, hasPlan, createdAt]
      properties:
        patientId:
          type: string
        userId:
          type: string
        displayName:
          type: string
        email:
          type: [string, 'null']
        conditionLabel:
          type: string
        painLevel:
          type: integer
        latestPainScore:
          type: [integer, 'null']
        latestAdherence:
          type: [integer, 'null']
        lastCheckinAt:
          type: [string, 'null']
          format: date-time
        hasPlan:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AdminPatientDetail:
      type: object
      required: [profile, user, plan, recentTracking, notes]
      properties:
        profile:
          $ref: '#/components/schemas/PatientProfile'
        user:
          $ref: '#/components/schemas/AuthUser'
        plan:
          oneOf:
            - $ref: '#/components/schemas/RecoveryPlan'
            - type: 'null'
        recentTracking:
          type: array
          items:
            $ref: '#/components/schemas/TrackingEntry'
        notes:
          type: array
          items:
            $ref: '#/components/schemas/ClinicianNote'
    ClinicianNote:
      type: object
      required: [id, patientId, authorId, authorName, content, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateClinicianNoteBody:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
    AdminConditionBreakdown:
      type: object
      required: [conditionLabel, count]
      properties:
        conditionLabel:
          type: string
        count:
          type: integer
    AdminDashboard:
      type: object
      required: [totalPatients, activePatientsLast7Days, averagePainScore, averageAdherence, conditionBreakdown, recentCheckins]
      properties:
        totalPatients:
          type: integer
        activePatientsLast7Days:
          type: integer
        averagePainScore:
          type: [number, 'null']
        averageAdherence:
          type: [number, 'null']
        conditionBreakdown:
          type: array
          items:
            $ref: '#/components/schemas/AdminConditionBreakdown'
        recentCheckins:
          type: array
          items:
            $ref: '#/components/schemas/AdminCheckinRow'
    AdminCheckinRow:
      type: object
      required: [patientId, displayName, conditionLabel, painScore, adherencePercent, date]
      properties:
        patientId:
          type: string
        displayName:
          type: string
        conditionLabel:
          type: string
        painScore:
          type: integer
        adherencePercent:
          type: integer
        date:
          type: string
          format: date