# lib/api-zod/src/generated/types/chatExchange.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecoveryOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface ChatExchange {
  userMessage: ChatMessage;
  assistantMessage: ChatMessage;
}