# lib/api-zod/src/generated/types/createClinicianNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecoveryOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClinicianNoteBody {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  content: string;
}