# lib/api-zod/src/generated/types/recoveryPlanContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecoveryOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";
import type { WeeklyDay } from "./weeklyDay";

export interface RecoveryPlanContent {
  summary: string;
  exercises: Exercise[];
  lifestyleTips: string[];
  precautions: string[];
  weeklyPlan: WeeklyDay[];
}