# lib/api-zod/src/generated/types/trackingEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecoveryOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrackingEntry {
  id: string;
  patientId: string;
  date: Date;
  painScore: number;
  completedExercises: number;
  totalExercises: number;
  adherencePercent: number;
  notes: string;
  createdAt: Date;
}