# lib/api-zod/src/generated/types/weeklyDay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RecoveryOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyDay {
  day: string;
  focus: string;
  activities: string[];
}